Modularize floating meeting overlays
Separate floating overlay settings normalization, route and transcript derivation, native panel I/O, and React lifecycle orchestration behind the existing host facade without changing behavior.
