+++ apps/desktop/src/meeting-float/host.tsx
@@ -4,7 +4,6 @@ import { useRef } from "react";
 import {
   commands as windowsCommands,
   events as windowsEvents,
-  type FloatingBarSettingsChange,
 } from "@anlg/plugin-windows";
 
 import {
@@ -13,120 +12,48 @@ import {
   type MeetingFloatData,
   subscribeMeetingFloatData,
 } from "./hooks";
+import {
+  getCurrentFloatingBarColorScheme,
+  getFloatingLiveCaptionToggleVisible,
+  getFloatingRouteState,
+  isSameFloatingRouteState,
+  type FloatingRouteState,
+  type ListenerState,
+} from "./route-state";
+import {
+  DEFAULT_FLOATING_OVERLAY_SETTINGS,
+  FLOATING_OVERLAY_SETTING_KEYS,
+  getFloatingOverlaySettings,
+  getSettingsValuesFromNativeChange,
+  type FloatingOverlaySettings,
+} from "./settings";
+import {
+  hideFloatingMeetingPanel,
+  hideLiveCaptionPanel,
+  showFloatingMeetingWindow,
+  syncFloatingMeetingWindow,
+} from "./window-panel";
 
 import {
   getStoredSettingValues,
   setSettingValue,
   useSetSettingValues,
 } from "~/settings/queries";
-import type { SettingValues } from "~/settings/schema";
 import { useConfigValue, useConfigValues } from "~/shared/config";
 import { useLatestRef } from "~/shared/hooks/useLatestRef";
 import { useMountEffect } from "~/shared/hooks/useMountEffect";
 import { listenerStore } from "~/store/zustand/listener/instance";
-import { SegmentKeyUtils, type RenderLabelContext } from "~/stt/live-segment";
+import type { RenderLabelContext } from "~/stt/live-segment";
 
-type ListenerState = ReturnType<typeof listenerStore.getState>;
-type FloatingBarStatus = "recording" | "error";
-type FloatingBarColorScheme = "light" | "dark";
-type LiveCaptionPosition =
-  | "topCenter"
-  | "topLeft"
-  | "topRight"
-  | "bottomLeft"
-  | "bottomRight"
-  | "bottomCenter";
-type FloatingOverlaySettings = {
-  floatingBarOpacity: number;
-  liveCaptionOpacity: number;
-  liveCaptionWidth: number;
-  liveCaptionLineCount: number;
-  liveCaptionPosition: LiveCaptionPosition;
-  liveCaptionMinimized: boolean;
-};
-type FloatingOverlaySettingsStorage = Pick<
-  SettingValues,
-  | "floating_bar_opacity"
-  | "live_caption_opacity"
-  | "live_caption_width"
-  | "live_caption_line_count"
-  | "live_caption_position"
-  | "live_caption_minimized"
->;
-type FloatingTranscriptBubble = {
-  id: string;
-  speakerLabel: string;
-  text: string;
-  isSelf: boolean;
-  isFinal: boolean;
-  startMs: number;
-  endMs: number;
-  overlapsPrevious: boolean;
-  overlapsNext: boolean;
-};
-type FloatingRouteState = {
-  sessionId: string;
-  title: string;
-  amplitude: number;
-  status: FloatingBarStatus;
-  colorScheme: FloatingBarColorScheme;
-  opacity: number;
-  liveCaptionOpacity: number;
-  liveCaptionWidth: number;
-  liveCaptionLineCount: number;
-  liveCaptionPosition: LiveCaptionPosition;
-  liveCaptionMinimized: boolean;
-  liveCaptionToggleVisible: boolean;
-  transcriptBubbles: FloatingTranscriptBubble[];
-};
-type LiveCaptionRouteState = {
-  sessionId: string;
-  text: string;
-  opacity: number;
-  width: number;
-  lineCount: number;
-  position: LiveCaptionPosition;
-  minimized: boolean;
-};
-
-const DEFAULT_FLOATING_OVERLAY_SETTINGS: FloatingOverlaySettings = {
-  floatingBarOpacity: 0.78,
-  liveCaptionOpacity: 0.3,
-  liveCaptionWidth: 440,
-  liveCaptionLineCount: 1,
-  liveCaptionPosition: "topCenter",
-  liveCaptionMinimized: true,
-};
-
-const FLOATING_BAR_MIN_OPACITY = 0.35;
-const FLOATING_BAR_MAX_OPACITY = 0.95;
-const LIVE_CAPTION_MIN_OPACITY = 0.05;
-const LIVE_CAPTION_MAX_OPACITY = 1;
-const LIVE_CAPTION_MIN_WIDTH = 260;
-const LIVE_CAPTION_MAX_WIDTH = 640;
-const LIVE_CAPTION_MIN_LINE_COUNT = 1;
-const LIVE_CAPTION_MAX_LINE_COUNT = 4;
-const LIVE_CAPTION_HORIZONTAL_PADDING_PX = 32;
-const LIVE_CAPTION_AVERAGE_CHARACTER_WIDTH_PX = 7.8;
-const FLOATING_TRANSCRIPT_OVERLAP_THRESHOLD_MS = 300;
-
-const LIVE_CAPTION_POSITIONS: ReadonlySet<string> = new Set([
-  "topCenter",
-  "topLeft",
-  "topRight",
-  "bottomLeft",
-  "bottomRight",
-  "bottomCenter",
-]);
-
-const FLOATING_OVERLAY_SETTING_KEYS = [
-  "floating_bar_opacity",
-  "live_caption_opacity",
-  "live_caption_width",
-  "live_caption_line_count",
-  "live_caption_position",
-  "live_caption_minimized",
-] as const;
+export {
+  getCurrentFloatingBarColorScheme,
+  getFloatingRouteState,
+  getFloatingTranscriptBubbles,
+  getLiveCaptionDisplayText,
+  getLiveCaptionRouteState,
+  shouldShowFloatingLiveCaptionToggle,
+} from "./route-state";
+export { hideFloatingMeetingPanel, hideLiveCaptionPanel } from "./window-panel";
 
 export function FloatingMeetingWindowHost() {
   const floatingBarEnabled = useConfigValue("floating_bar_enabled");
@@ -150,43 +77,6 @@ export function FloatingMeetingWindowHost() {
       <LiveCaptionWindowDisabled />
     </>
   );
-}
-
-function getFloatingOverlaySettings(
-  values: Partial<FloatingOverlaySettingsStorage>,
-): FloatingOverlaySettings {
-  return {
-    floatingBarOpacity: normalizeOpacity(
-      values.floating_bar_opacity,
-      DEFAULT_FLOATING_OVERLAY_SETTINGS.floatingBarOpacity,
-      FLOATING_BAR_MIN_OPACITY,
-      FLOATING_BAR_MAX_OPACITY,
-    ),
-    liveCaptionOpacity: normalizeOpacity(
-      values.live_caption_opacity,
-      DEFAULT_FLOATING_OVERLAY_SETTINGS.liveCaptionOpacity,
-      LIVE_CAPTION_MIN_OPACITY,
-      LIVE_CAPTION_MAX_OPACITY,
-    ),
-    liveCaptionWidth: normalizeNumber(
-      values.live_caption_width,
-      DEFAULT_FLOATING_OVERLAY_SETTINGS.liveCaptionWidth,
-      LIVE_CAPTION_MIN_WIDTH,
-      LIVE_CAPTION_MAX_WIDTH,
-    ),
-    liveCaptionLineCount: normalizeInteger(
-      values.live_caption_line_count,
-      DEFAULT_FLOATING_OVERLAY_SETTINGS.liveCaptionLineCount,
-      LIVE_CAPTION_MIN_LINE_COUNT,
-      LIVE_CAPTION_MAX_LINE_COUNT,
-    ),
-    liveCaptionPosition: normalizeLiveCaptionPosition(
-      values.live_caption_position,
-    ),
-    liveCaptionMinimized:
-      (values.live_caption_minimized ??
-        DEFAULT_FLOATING_OVERLAY_SETTINGS.liveCaptionMinimized) === true,
-  };
 }
 
 function FloatingOverlaySettingsEventSync() {
@@ -470,59 +360,6 @@ function FloatingMeetingWindowSettingsSync({
   return null;
 }
 
-export function getFloatingRouteState(
-  state: ListenerState,
-  {
-    sessionId,
-    colorScheme = "dark",
-    settings = DEFAULT_FLOATING_OVERLAY_SETTINGS,
-    liveCaptionToggleVisible = false,
-    sessionTitle,
-    speakerLabelContext,
-  }: {
-    sessionId?: string;
-    colorScheme?: FloatingBarColorScheme;
-    settings?: FloatingOverlaySettings;
-    liveCaptionToggleVisible?: boolean;
-    sessionTitle?: string | null;
-    speakerLabelContext?: RenderLabelContext;
-  } = {},
-): FloatingRouteState | null {
-  if (state.live.status !== "active") {
-    return null;
-  }
-
-  if (!state.live.sessionId) {
-    return null;
-  }
-
-  if (sessionId && state.live.sessionId !== sessionId) {
-    return null;
-  }
-
-  return {
-    sessionId: state.live.sessionId,
-    title: getFloatingTitle(sessionTitle),
-    amplitude: Math.min(
-      Math.hypot(state.live.amplitude.mic, state.live.amplitude.speaker),
-      1,
-    ),
-    status: state.live.degraded || state.live.lastError ? "error" : "recording",
-    colorScheme,
-    opacity: settings.floatingBarOpacity,
-    liveCaptionOpacity: settings.liveCaptionOpacity,
-    liveCaptionWidth: settings.liveCaptionWidth,
-    liveCaptionLineCount: settings.liveCaptionLineCount,
-    liveCaptionPosition: settings.liveCaptionPosition,
-    liveCaptionMinimized: settings.liveCaptionMinimized,
-    liveCaptionToggleVisible,
-    transcriptBubbles: getFloatingTranscriptBubbles(
-      state.liveSegments,
-      speakerLabelContext,
-    ),
-  };
-}
-
 function getCurrentFloatingRouteState(
   state: ListenerState,
   sessionId?: string,
@@ -552,89 +389,6 @@ function getFloatingSessionTitle(
   return meetingData?.sessions[sessionId]?.title ?? null;
 }
 
-function getFloatingTitle(title: string | null | undefined) {
-  const normalized = title?.trim();
-  return normalized || "Live transcript";
-}
-
-export function getFloatingTranscriptBubbles(
-  segments: ListenerState["liveSegments"],
-  speakerLabelContext?: RenderLabelContext,
-): FloatingTranscriptBubble[] {
-  const bubbles = segments
-    .slice()
-    .sort(
-      (a, b) =>
-        a.start_ms - b.start_ms ||
-        a.end_ms - b.end_ms ||
-        a.id.localeCompare(b.id),
-    )
-    .map((segment) => {
-      const text = getFloatingSegmentText(segment);
-      if (!text) {
-        return null;
-      }
-
-      return {
-        id: segment.id,
-        speakerLabel: getFloatingSpeakerLabel(segment.key, speakerLabelContext),
-        text,
-        isSelf: isFloatingSelfSpeaker(segment.key),
-        isFinal: segment.words.every((word) => word.is_final),
-        startMs: segment.start_ms,
-        endMs: segment.end_ms,
-        overlapsPrevious: false,
-        overlapsNext: false,
-      };
-    })
-    .filter((bubble): bubble is FloatingTranscriptBubble => bubble !== null);
-
-  return bubbles.map((bubble, index) => ({
-    ...bubble,
-    overlapsPrevious: bubbles
-      .slice(0, index)
-      .some((previous) => doFloatingTranscriptBubblesOverlap(previous, bubble)),
-    overlapsNext: bubbles
-      .slice(index + 1)
-      .some((next) => doFloatingTranscriptBubblesOverlap(bubble, next)),
-  }));
-}
-
-function getFloatingSegmentText(
-  segment: ListenerState["liveSegments"][number],
-) {
-  const wordText = segment.words
-    .map((word) => word.text.trim())
-    .filter(Boolean)
-    .join(" ")
-    .replace(/\s+([,.?!;:])/g, "$1");
-
-  return (wordText || segment.text).trim().replace(/\s+/g, " ");
-}
-
-function getFloatingSpeakerLabel(
-  key: ListenerState["liveSegments"][number]["key"],
-  ctx?: RenderLabelContext,
-) {
-  if (isFloatingSelfSpeaker(key)) {
-    return "You";
-  }
-
-  if (ctx) {
-    return SegmentKeyUtils.renderLabel(key, ctx);
-  }
-
-  if (key.speaker_index != null) {
-    return `Speaker ${key.speaker_index + 1}`;
-  }
-
-  if (key.channel === "RemoteParty") {
-    return "Speaker";
-  }
-
-  return "Audio";
-}
-
 function getFloatingSpeakerLabelContext(
   state: ListenerState,
   meetingData: MeetingFloatData | undefined,
@@ -644,77 +398,6 @@ function getFloatingSpeakerLabelContext(
   }
 
   return createMeetingFloatLabelContext(meetingData, state.live.sessionId);
-}
-
-function isFloatingSelfSpeaker(
-  key: ListenerState["liveSegments"][number]["key"],
-) {
-  return key.channel === "DirectMic";
-}
-
-function doFloatingTranscriptBubblesOverlap(
-  left: FloatingTranscriptBubble,
-  right: FloatingTranscriptBubble,
-) {
-  if (
-    left.isSelf === right.isSelf &&
-    left.speakerLabel === right.speakerLabel
-  ) {
-    return false;
-  }
-
-  const overlapMs =
-    Math.min(left.endMs, right.endMs) - Math.max(left.startMs, right.startMs);
-  return overlapMs >= FLOATING_TRANSCRIPT_OVERLAP_THRESHOLD_MS;
-}
-
-export function shouldShowFloatingLiveCaptionToggle({
-  liveTranscriptionActive,
-}: {
-  provider?: string | null;
-  model?: string | null;
-  liveTranscriptionActive: boolean;
-}) {
-  return liveTranscriptionActive;
-}
-
-function getFloatingLiveCaptionToggleVisible(state: ListenerState) {
-  return shouldShowFloatingLiveCaptionToggle({
-    liveTranscriptionActive: state.live.liveTranscriptionActive === true,
-  });
-}
-
-export function getLiveCaptionRouteState(
-  state: ListenerState,
-  settings: FloatingOverlaySettings = DEFAULT_FLOATING_OVERLAY_SETTINGS,
-): LiveCaptionRouteState | null {
-  if (state.live.status !== "active") {
-    return null;
-  }
-
-  if (!state.live.sessionId) {
-    return null;
-  }
-
-  if (state.live.liveTranscriptionActive !== true) {
-    return null;
-  }
-
-  if (settings.liveCaptionMinimized) {
-    return null;
-  }
-
-  const text = getLiveCaptionDisplayText(state.liveCaptionText, settings);
-
-  return {
-    sessionId: state.live.sessionId,
-    text,
-    opacity: settings.liveCaptionOpacity,
-    width: settings.liveCaptionWidth,
-    lineCount: settings.liveCaptionLineCount,
-    position: settings.liveCaptionPosition,
-    minimized: settings.liveCaptionMinimized,
-  };
 }
 
 function subscribeToAppliedTheme(onStoreChange: () => void) {
@@ -731,296 +414,6 @@ function subscribeToAppliedTheme(onStoreChange: () => void) {
     attributes: true,
   });
   return () => observer.disconnect();
-}
-
-export function getCurrentFloatingBarColorScheme(): FloatingBarColorScheme {
-  if (typeof document === "undefined") {
-    return "dark";
-  }
-
-  return document.documentElement.classList.contains("dark") ? "dark" : "light";
-}
-
-function isSameFloatingRouteState(
-  left: FloatingRouteState | null,
-  right: FloatingRouteState | null,
-) {
-  return (
-    left?.sessionId === right?.sessionId &&
-    left?.amplitude === right?.amplitude &&
-    left?.status === right?.status &&
-    left?.colorScheme === right?.colorScheme &&
-    left?.opacity === right?.opacity &&
-    left?.liveCaptionOpacity === right?.liveCaptionOpacity &&
-    left?.liveCaptionWidth === right?.liveCaptionWidth &&
-    left?.liveCaptionLineCount === right?.liveCaptionLineCount &&
-    left?.liveCaptionPosition === right?.liveCaptionPosition &&
-    left?.liveCaptionMinimized === right?.liveCaptionMinimized &&
-    left?.liveCaptionToggleVisible === right?.liveCaptionToggleVisible &&
-    left?.title === right?.title &&
-    isSameFloatingTranscriptBubbles(
-      left?.transcriptBubbles,
-      right?.transcriptBubbles,
-    )
-  );
-}
-
-function isSameFloatingTranscriptBubbles(
-  left: FloatingTranscriptBubble[] | undefined,
-  right: FloatingTranscriptBubble[] | undefined,
-) {
-  if (left === right) {
-    return true;
-  }
-
-  if (!left || !right || left.length !== right.length) {
-    return false;
-  }
-
-  return left.every((bubble, index) => {
-    const other = right[index];
-    return (
-      other &&
-      bubble.id === other.id &&
-      bubble.speakerLabel === other.speakerLabel &&
-      bubble.text === other.text &&
-      bubble.isSelf === other.isSelf &&
-      bubble.isFinal === other.isFinal &&
-      bubble.startMs === other.startMs &&
-      bubble.endMs === other.endMs &&
-      bubble.overlapsPrevious === other.overlapsPrevious &&
-      bubble.overlapsNext === other.overlapsNext
-    );
-  });
-}
-
-async function syncFloatingMeetingWindow(
-  routeState: FloatingRouteState | null,
-  shownSessionId: string | null,
-  shouldContinue: () => boolean,
-): Promise<string | null | "unavailable"> {
-  if (!shouldContinue()) {
-    return null;
-  }
-
-  if (!routeState) {
-    await hideFloatingMeetingPanel();
-    return null;
-  }
-
-  const ready = await showFloatingMeetingWindow(
-    routeState,
-    shownSessionId !== routeState.sessionId,
-    shouldContinue,
-  );
-  if (!shouldContinue()) {
-    await hideFloatingMeetingPanel();
-    return null;
-  }
-
-  return ready ? routeState.sessionId : "unavailable";
-}
-
-async function showFloatingMeetingWindow(
-  routeState: FloatingRouteState,
-  shouldShow: boolean,
-  shouldContinue: () => boolean = () => true,
-): Promise<boolean> {
-  if (!shouldContinue()) {
-    return false;
-  }
-
-  if (shouldShow) {
-    const showResult = await windowsCommands.floatingBarShow();
-    if (!shouldContinue()) {
-      await hideFloatingMeetingPanel();
-      return false;
-    }
-
-    if (showResult.status === "error") {
-      console.error("Failed to show floating meeting panel:", showResult.error);
-      return false;
-    }
-  }
-
-  const updateResult = await windowsCommands.floatingBarUpdate({
-    amplitude: routeState.amplitude,
-    title: routeState.title,
-    status: routeState.status,
-    colorScheme: routeState.colorScheme,
-    opacity: routeState.opacity,
-    liveCaptionOpacity: routeState.liveCaptionOpacity,
-    liveCaptionWidth: routeState.liveCaptionWidth,
-    liveCaptionLineCount: routeState.liveCaptionLineCount,
-    liveCaptionPosition: routeState.liveCaptionPosition,
-    liveCaptionMinimized: routeState.liveCaptionMinimized,
-    liveCaptionToggleVisible: routeState.liveCaptionToggleVisible,
-    transcriptBubbles: routeState.transcriptBubbles,
-  });
-  if (!shouldContinue()) {
-    await hideFloatingMeetingPanel();
-    return false;
-  }
-
-  if (updateResult.status === "error") {
-    console.error(
-      "Failed to update floating meeting panel:",
-      updateResult.error,
-    );
-    return false;
-  }
-
-  return true;
-}
-
-export function getLiveCaptionDisplayText(
-  text: string,
-  settings: Pick<
-    FloatingOverlaySettings,
-    "liveCaptionWidth" | "liveCaptionLineCount"
-  > = DEFAULT_FLOATING_OVERLAY_SETTINGS,
-) {
-  const normalizedText = text.trim().replace(/\s+/g, " ");
-  if (!normalizedText) {
-    return "";
-  }
-
-  const contentWidth = Math.max(
-    settings.liveCaptionWidth - LIVE_CAPTION_HORIZONTAL_PADDING_PX,
-    LIVE_CAPTION_MIN_WIDTH - LIVE_CAPTION_HORIZONTAL_PADDING_PX,
-  );
-  const charactersPerLine = Math.max(
-    12,
-    Math.floor(contentWidth / LIVE_CAPTION_AVERAGE_CHARACTER_WIDTH_PX),
-  );
-  const maxCharacters = Math.max(
-    24,
-    charactersPerLine * settings.liveCaptionLineCount,
-  );
-
-  if (normalizedText.length <= maxCharacters) {
-    return normalizedText;
-  }
-
-  return `... ${getTextSuffixAtWordBoundary(normalizedText, maxCharacters - 4)}`;
-}
-
-function getTextSuffixAtWordBoundary(text: string, maxCharacters: number) {
-  const suffix = text.slice(-maxCharacters).trimStart();
-  const firstWhitespaceIndex = suffix.search(/\s/);
-  if (firstWhitespaceIndex > 0 && firstWhitespaceIndex < suffix.length - 1) {
-    return suffix.slice(firstWhitespaceIndex).trimStart();
-  }
-
-  return suffix;
-}
-function normalizeOpacity(
-  value: unknown,
-  fallback: number,
-  min: number,
-  max: number,
-): number {
-  return normalizeNumber(value, fallback, min, max);
-}
-
-function normalizeNumber(
-  value: unknown,
-  fallback: number,
-  min: number,
-  max: number,
-): number {
-  if (typeof value !== "number" || !Number.isFinite(value)) {
-    return fallback;
-  }
-
-  return Math.min(Math.max(value, min), max);
-}
-
-function normalizeInteger(
-  value: unknown,
-  fallback: number,
-  min: number,
-  max: number,
-): number {
-  return Math.round(normalizeNumber(value, fallback, min, max));
-}
-
-function normalizeLiveCaptionPosition(value: unknown): LiveCaptionPosition {
-  if (typeof value === "string" && LIVE_CAPTION_POSITIONS.has(value)) {
-    return value as LiveCaptionPosition;
-  }
-
-  return DEFAULT_FLOATING_OVERLAY_SETTINGS.liveCaptionPosition;
-}
-
-function getSettingsValuesFromNativeChange(change: FloatingBarSettingsChange) {
-  const values: Partial<FloatingOverlaySettingsStorage> = {};
-
-  if (
-    change.floatingBarOpacity !== null &&
-    change.floatingBarOpacity !== undefined
-  ) {
-    values.floating_bar_opacity = normalizeOpacity(
-      change.floatingBarOpacity,
-      DEFAULT_FLOATING_OVERLAY_SETTINGS.floatingBarOpacity,
-      FLOATING_BAR_MIN_OPACITY,
-      FLOATING_BAR_MAX_OPACITY,
-    );
-  }
-
-  if (
-    change.liveCaptionOpacity !== null &&
-    change.liveCaptionOpacity !== undefined
-  ) {
-    values.live_caption_opacity = normalizeOpacity(
-      change.liveCaptionOpacity,
-      DEFAULT_FLOATING_OVERLAY_SETTINGS.liveCaptionOpacity,
-      LIVE_CAPTION_MIN_OPACITY,
-      LIVE_CAPTION_MAX_OPACITY,
-    );
-  }
-
-  if (
-    change.liveCaptionWidth !== null &&
-    change.liveCaptionWidth !== undefined
-  ) {
-    values.live_caption_width = normalizeNumber(
-      change.liveCaptionWidth,
-      DEFAULT_FLOATING_OVERLAY_SETTINGS.liveCaptionWidth,
-      LIVE_CAPTION_MIN_WIDTH,
-      LIVE_CAPTION_MAX_WIDTH,
-    );
-  }
-
-  if (
-    change.liveCaptionLineCount !== null &&
-    change.liveCaptionLineCount !== undefined
-  ) {
-    values.live_caption_line_count = normalizeInteger(
-      change.liveCaptionLineCount,
-      DEFAULT_FLOATING_OVERLAY_SETTINGS.liveCaptionLineCount,
-      LIVE_CAPTION_MIN_LINE_COUNT,
-      LIVE_CAPTION_MAX_LINE_COUNT,
-    );
-  }
-
-  if (
-    change.liveCaptionPosition !== null &&
-    change.liveCaptionPosition !== undefined
-  ) {
-    values.live_caption_position = normalizeLiveCaptionPosition(
-      change.liveCaptionPosition,
-    );
-  }
-
-  if (
-    change.liveCaptionMinimized !== null &&
-    change.liveCaptionMinimized !== undefined
-  ) {
-    values.live_caption_minimized = change.liveCaptionMinimized === true;
-  }
-
-  return values;
 }
 
 export async function openFloatingMeetingPanel({
@@ -1053,18 +446,4 @@ export async function openFloatingMeetingPanel({
   }
 
   await showFloatingMeetingWindow(routeState, true);
-}
-
-export async function hideFloatingMeetingPanel() {
-  const result = await windowsCommands.floatingBarHide();
-  if (result.status === "error") {
-    console.error("Failed to hide floating meeting panel:", result.error);
-  }
-}
-
-export async function hideLiveCaptionPanel() {
-  const result = await windowsCommands.liveCaptionHide();
-  if (result.status === "error") {
-    console.error("Failed to hide live caption panel:", result.error);
-  }
 }
