+++ apps/desktop/src/meeting-float/window-panel.ts
@@ -1,0 +1,96 @@
+import { commands as windowsCommands } from "@anlg/plugin-windows";
+
+import type { FloatingRouteState } from "./route-state";
+
+export async function syncFloatingMeetingWindow(
+  routeState: FloatingRouteState | null,
+  shownSessionId: string | null,
+  shouldContinue: () => boolean,
+): Promise<string | null | "unavailable"> {
+  if (!shouldContinue()) {
+    return null;
+  }
+
+  if (!routeState) {
+    await hideFloatingMeetingPanel();
+    return null;
+  }
+
+  const ready = await showFloatingMeetingWindow(
+    routeState,
+    shownSessionId !== routeState.sessionId,
+    shouldContinue,
+  );
+  if (!shouldContinue()) {
+    await hideFloatingMeetingPanel();
+    return null;
+  }
+
+  return ready ? routeState.sessionId : "unavailable";
+}
+
+export async function showFloatingMeetingWindow(
+  routeState: FloatingRouteState,
+  shouldShow: boolean,
+  shouldContinue: () => boolean = () => true,
+): Promise<boolean> {
+  if (!shouldContinue()) {
+    return false;
+  }
+
+  if (shouldShow) {
+    const showResult = await windowsCommands.floatingBarShow();
+    if (!shouldContinue()) {
+      await hideFloatingMeetingPanel();
+      return false;
+    }
+
+    if (showResult.status === "error") {
+      console.error("Failed to show floating meeting panel:", showResult.error);
+      return false;
+    }
+  }
+
+  const updateResult = await windowsCommands.floatingBarUpdate({
+    amplitude: routeState.amplitude,
+    title: routeState.title,
+    status: routeState.status,
+    colorScheme: routeState.colorScheme,
+    opacity: routeState.opacity,
+    liveCaptionOpacity: routeState.liveCaptionOpacity,
+    liveCaptionWidth: routeState.liveCaptionWidth,
+    liveCaptionLineCount: routeState.liveCaptionLineCount,
+    liveCaptionPosition: routeState.liveCaptionPosition,
+    liveCaptionMinimized: routeState.liveCaptionMinimized,
+    liveCaptionToggleVisible: routeState.liveCaptionToggleVisible,
+    transcriptBubbles: routeState.transcriptBubbles,
+  });
+  if (!shouldContinue()) {
+    await hideFloatingMeetingPanel();
+    return false;
+  }
+
+  if (updateResult.status === "error") {
+    console.error(
+      "Failed to update floating meeting panel:",
+      updateResult.error,
+    );
+    return false;
+  }
+
+  return true;
+}
+
+export async function hideFloatingMeetingPanel() {
+  const result = await windowsCommands.floatingBarHide();
+  if (result.status === "error") {
+    console.error("Failed to hide floating meeting panel:", result.error);
+  }
+}
+
+export async function hideLiveCaptionPanel() {
+  const result = await windowsCommands.liveCaptionHide();
+  if (result.status === "error") {
+    console.error("Failed to hide live caption panel:", result.error);
+  }
+}
